refactor: add feature directory
